+++ src/ee/voltdbipc.cpp
@@ -131,7 +131,7 @@ public:
 
     int64_t getQueuedExportBytes(int32_t partitionId, std::string signature);
     void pushExportBuffer(int64_t exportGeneration, int32_t partitionId, std::string signature, voltdb::StreamBlock *block, bool sync, bool endOfStream);
-    int reportDRConflict(int32_t partitionId,
+    int reportDRConflict(int64_t partitionId,
                 int64_t remoteSequenceNumber, voltdb::DRConflictType conflict_type, voltdb::DRRecordType action_type,
                 std::string tableName, voltdb::Table* existingTable, voltdb::Table* expectedTable,
                 voltdb::Table* newTable, voltdb::Table* output);
@@ -1561,7 +1561,7 @@ int64_t VoltDBIPC::pushDRBuffer(int32_t partitionId, voltdb::StreamBlock *block)
     return -1;
 }
 
-int VoltDBIPC::reportDRConflict(int32_t partitionId,
+int VoltDBIPC::reportDRConflict(int64_t partitionId,
         int64_t remoteSequenceNumber, voltdb::DRConflictType conflict_type,
         voltdb::DRRecordType action_type, std::string tableName,
         voltdb::Table* exisitingTable, voltdb::Table* expectedTable,
